Extract OpenConfigurator from Configurators namespace
