--- sig/kangaru/configurators/open_configurator.rbs
@@ -1,11 +1,0 @@
-module Kangaru
-  module Configurators
-    class OpenConfigurator < Configurator
-      def self.from_yaml_file: (String) -> instance
-
-      private
-
-      def set_accessors!: (**untyped) -> void
-    end
-  end
-end
